requirements.txt
numpy
scipy
soundfile
openai-whisper
anthropic
python-dotenv
colorama
